* BlamLib: Added support for 'OldStringId' in CheApe (just a tag_string field, with the definition set to 'sidy')
* YeloLib: Added winapi_handle_closer and crt_file_closer deleter types (for use with std::unique_ptr)
* Halo1: A few changes for string_id_yelo stuff
* Halo1: Started working on support for non-standard string fields (with lengths greater than 31)
* Halo1: Refactored s_object_data's scenario_location to just 'location'
* Halo1: Added more object_function definitions and documentation
* Halo1: Corrected K_REAL_EPSILON's definition
* Halo1_CE: corrected case statements in PredictMemoryPoolUsage
* Halo1_CheApe: CheApeCore now has range based for loops where possible
* Halo1_CheApe: Moved definitions in Engine/EngineFunctions.hpp to their respective blamlib files
* Halo1_CheApe: Added more extendable field_dialog extension support to YeloGuerilla.cpp
* Halo1_CheApe: Started support for allowing the string_field_dialog to editor non-standard string fields (eg, long_strings)

* BlamLib: TODO: CheApe compiler still needs to have non-standard string field support added

+++ OpenSauce/Halo2/Halo2_CheApe/Memory/_EngineLayout.OpenSauce.inl
@@ -15,7 +15,7 @@
 #elif __EL_INCLUDE_FILE_ID == __EL_MEMORY_CHE_APE_INTERFACE
 	namespace CheApe
 	{
-		static const uint32 TagGroupDefinitionsRefs[] = { // &tag_groups[x]
+		static const uintptr_t TagGroupDefinitionsRefs[] = { // &tag_groups[x]
 			PLATFORM_VALUE(0x482E06, 0x52D416, NULL),	PLATFORM_VALUE(0x482E29, 0x52D439, NULL),	PLATFORM_VALUE(0x482E52, 0x52D462, NULL),
 			PLATFORM_VALUE(0x483B62, 0x52E1C2, NULL),	PLATFORM_VALUE(0x483C07, 0x52E267, NULL),	PLATFORM_VALUE(0x484029, 0x52E689, NULL),
 			PLATFORM_VALUE(0x484051, 0x52E6B1, NULL),	PLATFORM_VALUE(0x486622, 0x530F52, NULL),	PLATFORM_VALUE(0x486684, 0x530FB4, NULL),
@@ -23,7 +23,7 @@
 			PLATFORM_VALUE(0x489B61, 0x5346C1, NULL),
 		};
 
-		static const uint32 TagGroupDefinitionsCountRefs8bit[] = { // CMP si, 77h
+		static const uintptr_t TagGroupDefinitionsCountRefs8bit[] = { // CMP si, 77h
 			PLATFORM_VALUE(0x482E12, 0x52D422, NULL),
 		};
