+++ OpenSauce/shared/Include/YeloLib/Halo1/tag_files/tag_fieldset_replacement_builder.hpp
@@ -11,7 +11,7 @@ namespace Yelo
 {
 	namespace TagGroups
 	{
-		template<typename TFieldSetType, int kMaxFieldCount>
+		template<typename TFieldSetType, size_t kMaxFieldCount>
 		class c_tag_fieldset_replacement_memory
 		{
 			static int g_new_fields_count;
@@ -29,13 +29,13 @@ namespace Yelo
 		class c_tag_fieldset_replacement_builder
 		{
 			const tag_field* m_source_fields;
-			int m_source_fields_cursor;
+			size_t m_source_fields_cursor;
 
 			tag_field* m_target_fields;
-			int m_target_fields_cursor;
-			int m_target_fields_max_count;
+			size_t m_target_fields_cursor;
+			size_t m_target_fields_max_count;
 
-			int m_available_pad_size;
+			size_t m_available_pad_size;
 			c_tag_field_scanner scanner;	// source_fields scanner, for field size/offset interop
 
 		private:
@@ -45,21 +45,22 @@ namespace Yelo
 
 			c_tag_fieldset_replacement_builder(
 				const tag_field* source_fields,
-				tag_field* target_fields, int target_fields_size, int target_fields_count);
+				tag_field* target_fields, size_t target_fields_size, size_t target_fields_count);
 		public:
-			template<typename TFieldSetType, int kMaxFieldCount>
+			template<typename TFieldSetType, size_t kMaxFieldCount>
 			c_tag_fieldset_replacement_builder(const tag_field* source_fields,
 				c_tag_fieldset_replacement_memory<TFieldSetType, kMaxFieldCount> memory)
 			{
+				// TODO: have to wait until VS2013 for delegating constructors (this is creating a temp object, not ctoring this)
 				c_tag_fieldset_replacement_builder(source_fields, 
 					memory.GetFields(), memory.GetMaxFieldCount(), memory.GetFieldCount())
 			}
 
-			void CopyFieldsFromSource(int field_count);
-			void SkipFieldsFromSource(int field_count);
+			void CopyFieldsFromSource(size_t field_count);
+			void SkipFieldsFromSource(size_t field_count);
 
-			void UselessPaddingBegin(int pad_size);
-			int UselessPaddingEnd();
+			void UselessPaddingBegin(size_t pad_size);
+			size_t UselessPaddingEnd();
 
 			tag_field& InsertField(tag_field insert_field, size_t expected_offset);
 		};
